# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PRD Sprint Planner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FeatureStatus = (typeof FeatureStatus)[keyof typeof FeatureStatus];

export const FeatureStatus = {
  pending: "pending",
  generating: "generating",
  completed: "completed",
  failed: "failed",
} as const;

export interface Feature {
  id: number;
  title: string;
  description: string;
  status: FeatureStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFeatureBody {
  title: string;
  description: string;
}

export interface Prd {
  id: number;
  featureId: number;
  featureTitle: string;
  overview: string;
  goals: string;
  userStories: string;
  technicalRequirements: string;
  successMetrics: string;
  totalTasks: number;
  totalSprints: number;
  totalEffortPoints: number;
  riskScore: number;
  createdAt: string;
  updatedAt: string;
}

export type TaskType = (typeof TaskType)[keyof typeof TaskType];

export const TaskType = {
  user_story: "user_story",
  engineering: "engineering",
  bug: "bug",
  infrastructure: "infrastructure",
  design: "design",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type TaskRiskLevel = (typeof TaskRiskLevel)[keyof typeof TaskRiskLevel];

export const TaskRiskLevel = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  backlog: "backlog",
  in_sprint: "in_sprint",
  in_progress: "in_progress",
  done: "done",
} as const;

export interface Task {
  id: number;
  prdId: number;
  /** @nullable */
  sprintId?: number | null;
  title: string;
  description: string;
  type: TaskType;
  priority: TaskPriority;
  effortPoints: number;
  priorityScore: number;
  riskLevel: TaskRiskLevel;
  dependencies: number[];
  status: TaskStatus;
  createdAt: string;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  backlog: "backlog",
  in_sprint: "in_sprint",
  in_progress: "in_progress",
  done: "done",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface UpdateTaskBody {
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  sprintId?: number | null;
}

export interface Sprint {
  id: number;
  prdId: number;
  sprintNumber: number;
  name: string;
  goal: string;
  totalEffortPoints: number;
  taskCount: number;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  createdAt: string;
}

export type AdminStatsTasksByPriority = {
  critical: number;
  high: number;
  medium: number;
  low: number;
};

export type AdminStatsTasksByType = {
  user_story: number;
  engineering: number;
  bug: number;
  infrastructure: number;
  design: number;
};

export interface AdminStats {
  totalFeatures: number;
  totalPrds: number;
  totalTasks: number;
  totalSprints: number;
  avgTasksPerPrd: number;
  avgSprintsPerPrd: number;
  avgEffortPerPrd: number;
  tasksByPriority: AdminStatsTasksByPriority;
  tasksByType: AdminStatsTasksByType;
  recentGenerations: number;
}

export interface ActivityItem {
  id: number;
  featureTitle: string;
  /** @nullable */
  prdId?: number | null;
  /** @nullable */
  taskCount?: number | null;
  /** @nullable */
  sprintCount?: number | null;
  status: string;
  createdAt: string;
}

export interface GeneratePrdBody {
  featureId: number;
}

export interface GeneratePrdResponse {
  prd: Prd;
  tasks: Task[];
  sprints: Sprint[];
}

export type ListTasksParams = {
  /**
   * @nullable
   */
  prdId?: number | null;
  /**
   * @nullable
   */
  sprintId?: number | null;
};

export type ListSprintsParams = {
  /**
   * @nullable
   */
  prdId?: number | null;
};